Add usePatchAccessToken
