+++ packages/api/common/src/hooks/auth/index.ts
@@ -1,0 +1,1 @@
+export * from './usePatchAccessToken';
