+++ packages/api/common/src/hooks/index.ts
@@ -1,2 +1,2 @@
-// export * from './auth';
 export * from './mission';
+export * from './auth';
