+++ packages/api/common/src/libs/queryKeys.ts
@@ -1,4 +1,8 @@
 export const missionQueryKeys = {
-  getMissionList: () => ['unapprove', 'list'],
+  getMissionList: () => ['mission', 'list'],
   getMissionDetail: (missionId: string) => ['mission', 'detail', missionId],
 } as const;
+
+export const refreshQueryKeys = {
+  patchAccessToken: () => ['token', 'accessToken'],
+};
